Add a typing module with a decorator to disable `__init__`
This commit adds a new module `frequenz.core.typing` with a decorator
`disable_init` that can be used to disable the `__init__` constructor of
a class. This is useful when a class doesn't provide a default
constructor and requires the use of a factory method to create
instances.

The code is loosely based on the `_NoDefaultConstructible` class in the
SDK [`frequenz.sdk.timeseries._quantities`][1], but it was completely
rewritten to be more generic and to use a decorator instead of
inheritance, which makes also much easier to use.

```python
@disable_init
class MyClass(Super):
    pass
```

Versus:

```python
class MyClass(
    Super,
    metaclass=_NoDefaultConstructible,
):
    pass
```

Also it is safer to use, as it will raise an error as soon as a class is
declared with an `__init__` method.

[1]: https://github.com/frequenz-floss/frequenz-sdk-python/blob/88bd28e2f21b930279d66c1daf3deeaf6f9bb07a/src/frequenz/sdk/timeseries/_quantities.py#L505-L522

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>