Rename `BackgroundService`'s `name` to `unique_id`
Using `name` was confusing, as people didn't realize that it is used to
actually identify one of multiple instances. Also people were very prone
to use the class name as `name`, which is redundant as the class name
is also included in `str` and `repr`.

Using `unique_id` makes it more clear that this is used for
identification in logs and that a default is better than passing a
string that is not really unique (like the class name).

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>