Use the task creator object to create tasks in the Service class
This change adds a new parameter to the Service class, `task_creator`,
that will be used to create tasks. This is useful to allow the user to
use different task creators, like the `asyncio` module or a `TaskGroup`.

A method to create tasks was also added to the Service class, that uses
the task creator object to create the tasks, hopefully making it easier
for users not to forget to register tasks managed by the service in the
`self._tasks` set.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>