Add an option to log exceptions raised by tasks
It is common that users forget to handle exceptions properly in their
tasks, and since services tasks are supposed to run forever, there are
no many opportunities to await for their completion appropriately and
handle the failure.

With this at least we make sure we will get a log message when a task
raises an exception, so we can at least know what happened and they just
silently disappear.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>