Make `Service` an ABC and add a `ServiceBase`
By having both mixed up, it was hard to separate documentation for users
of services, and developers wanting to implement services. It also gave
access to service users to internals for service implementation, like
`create_task`, which is not ideal.

Now users can use the `Service` type to refer to services generically
and implementors can use `ServiceBase` to have a sane starting point to
implement a service.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>