Add a `PersistentTaskGroup` class
`asyncio.TaskGroup` is a very convenient construct when using
parallelization for doing calculations for example, where the results
for all the tasks need to be merged together to produce a final result.
In this case if one of the tasks fails, it makes sense to cancel the
others and abort as soon as possible, as any further calculations would
be thrown away.

This class is intended to help managing a group of tasks that should
persist even if other tasks in the group fail, usually by either only
discarding the failed task or by restarting it somehow.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>